Application service refactor
- Removed unused method
- Changed request classes import packages
